Most of the Wall portion of the Dungeon script is complete
Aside from some random bug I can't figure out, I'll note it here and hit Chris up in the group chat. The one bug seemingly random portions of the walls, which don't generate and I can't find a pattern to the ones that don't. Additionally, the placement is still messed up as the walls all spawn signifcantly to the left of the map at the time of generation.